requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncer==0.0.7
bidict==0.23.1
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.5
filetype==1.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.0
jmespath==1.0.1
Lazify==0.4.0
literalai==0.0.623
marshmallow==3.23.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.55.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==23.2
protobuf==5.29.0
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
requests==2.32.3
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
syncer==2.0.3
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.28.2
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
